winter cleaning
- implement error handler for custom 404 pages
- move blog -> posts
- fix bold font loading
- stamp stylesheet queries
- make git hard reset only happen when in container

+++ on_hook.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
-BRANCH=$(git rev-parse --abbrev-ref HEAD)
-git fetch origin && git reset --hard origin/$BRANCH
-./www/blog/update.sh
+set -xv
+if [[ ! -z "${DEFINITELY_RUNNING_IN_PRODUCTION}" ]]; then
+    BRANCH=$(git rev-parse --abbrev-ref HEAD)
+    git fetch origin && git reset --hard origin/$BRANCH
+fi
+./www/posts/update.sh
